whitespace

+++ include/leap/hash.h
@@ -1,10 +1,6 @@
 //
 // hash.h
 //
-// hash routines
-//
-//   Peter Niekamp, February, 2014
-//
 
 #ifndef HASHLIB_HH
 #define HASHLIB_HH
@@ -23,6 +19,7 @@ namespace leap { namespace crypto
 
   //-------------------------- MD5 --------------------------------------------
   //---------------------------------------------------------------------------
+
   struct md5
   {
     md5();
@@ -40,14 +37,15 @@ namespace leap { namespace crypto
   };
 
   void md5_init(md5 *context);
-  void md5_update(md5 *context, const void *data, unsigned long size);
+  void md5_update(md5 *context, const void *data, size_t size);
   void md5_finalise(md5 *context);
 
-  md5 md5digest(const void *data, unsigned long size);
+  md5 md5digest(const void *data, size_t size);
 
 
   //-------------------------- SHA1 -------------------------------------------
   //---------------------------------------------------------------------------
+
   struct sha1
   {
     sha1();
@@ -65,10 +63,10 @@ namespace leap { namespace crypto
   };
 
   void sha1_init(sha1 *context);
-  void sha1_update(sha1 *context, const void *data, unsigned long size);
+  void sha1_update(sha1 *context, const void *data, size_t size);
   void sha1_finalise(sha1 *context);
 
-  sha1 sha1digest(const void *data, unsigned long size);
+  sha1 sha1digest(const void *data, size_t size);
 
 } } // namespace crypto
